Added 2 new enemy types and updated the enemy spawner
Added a dashing enemy and a shooting enemy and made the auto fire mechanic interact with them. Also added them into the enemy pool for the spawner to take from so they randomly spawn like the basic enemy. Removed enemy-specific tags as they were not doing anything and only made it harder for the player to interact with them.